refactored pdfshare + added pdfbsd(pdf book scanner daemon)
