+++ pdfshare/app/models/book_db_client.py
@@ -1,4 +1,4 @@
-from extensions.book_handler import BookHandler
+from pdfshare.app.extensions.book_handler import BookHandler
 
 # create instances of bookdb
 books_db_client = BookHandler()
