+++ pdfshare/app/static/books/.gitignore
