+++ pdfshare/app/static/covers/.gitignore
