+++ pdfshare/app/static/css/config-box.css
