+++ pdfshare/app/static/css/gallery-view.css
