+++ pdfshare/app/static/js/config-box.js
