+++ pdfshare/app/static/js/gallery-view.js
