+++ pdfshare/app/static/js/list-view.js
