+++ pdfshare/app/templates/gallery-view.jinja2
