+++ pdfshare/app/wsgi.py
@@ -1,5 +1,4 @@
-from main import app
-
+from pdfshare.app.main import app
 
 if __name__ == "__main__":
     app.run()
