+++ pdfshare/common/clients/book_client.py
