+++ requirements.txt
@@ -2,3 +2,5 @@ Flask==1.1.1
 gunicorn==19.9.0
 watchdog==2.1.0
 pymongo==3.10.1
+watchdog==2.1.0
+python-daemon==2.3.0
